move SanitizeName
